# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Paadi backend — auth, pots, payments, settlements",
    "title": "Paadi API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/me/api-keys": {
      "get": {
        "description": "Returns the caller's API keys (prefix and metadata only — never the secret). Session auth only.",
        "operationId": "ApiKeysController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "keys": {
                      "items": {
                        "properties": {
                          "createdAt": {
                            "type": "string"
                          },
                          "expiresAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "lastUsedAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "prefix": {
                            "type": "string"
                          },
                          "revokedAt": {
                            "nullable": true,
                            "type": "string"
                          },
                          "scopes": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "prefix",
                          "scopes",
                          "lastUsedAt",
                          "expiresAt",
                          "revokedAt",
                          "createdAt",
                          "updatedAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "keys"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List API keys",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "Creates a scoped API key acting as the caller. Session auth only — the plaintext key is returned once in this response and never again.",
        "operationId": "ApiKeysController_mint",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "expiresAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "name": {
                    "maxLength": 64,
                    "minLength": 1,
                    "type": "string"
                  },
                  "scopes": {
                    "items": {
                      "enum": [
                        "pots:read",
                        "pots:write",
                        "wallet:read",
                        "wallet:pay",
                        "wallet:withdraw",
                        "bills:read",
                        "profile:read",
                        "activity:read",
                        "webhooks:manage"
                      ],
                      "type": "string"
                    },
                    "minItems": 1,
                    "type": "array"
                  }
                },
                "required": [
                  "name",
                  "scopes"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "lastUsedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "prefix": {
                      "type": "string"
                    },
                    "revokedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "scopes": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "prefix",
                    "scopes",
                    "lastUsedAt",
                    "expiresAt",
                    "revokedAt",
                    "createdAt",
                    "updatedAt",
                    "key"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mint API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/me/api-keys/current": {
      "get": {
        "description": "Returns the identity, mode and scopes of the API key making this call. API-key auth only — lets an integration discover what it is allowed to do.",
        "operationId": "ApiKeysController_current",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "mode": {
                      "enum": [
                        "live",
                        "test"
                      ],
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "prefix": {
                      "type": "string"
                    },
                    "scopes": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "prefix",
                    "mode",
                    "scopes"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Introspect current API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/me/api-keys/{id}": {
      "delete": {
        "description": "Revokes an API key the caller owns. Takes effect immediately. Session auth only.",
        "operationId": "ApiKeysController_revoke",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "lastUsedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "prefix": {
                      "type": "string"
                    },
                    "revokedAt": {
                      "nullable": true,
                      "type": "string"
                    },
                    "scopes": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "prefix",
                    "scopes",
                    "lastUsedAt",
                    "expiresAt",
                    "revokedAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke API key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
  "servers": [],
...
}